+++ Chirp.sln
@@ -25,6 +25,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleDBTest", "test\Simple
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.CSVDBService", "src\Chirp.CSVDBService\Chirp.CSVDBService.csproj", "{C9367545-CF43-4709-A7D7-EFE620E0C0E3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Chirp.Core", "src\Chirp.Core\Chirp.Core.csproj", "{046F33B0-9573-4AFA-9A68-9B16344F9AF7}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -51,6 +53,10 @@ Global
 		{C9367545-CF43-4709-A7D7-EFE620E0C0E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C9367545-CF43-4709-A7D7-EFE620E0C0E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{C9367545-CF43-4709-A7D7-EFE620E0C0E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{046F33B0-9573-4AFA-9A68-9B16344F9AF7}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -61,6 +67,7 @@ Global
 		{FB3065B1-5801-4C04-9E1E-6C6422072F8E} = {AAF0C3E6-2C01-4C84-9E9D-77812EF704AD}
 		{EC69DBBF-7546-406B-980E-C6EA83A3CBEA} = {AAF0C3E6-2C01-4C84-9E9D-77812EF704AD}
 		{C9367545-CF43-4709-A7D7-EFE620E0C0E3} = {8F8A6A34-59A8-446F-A994-A98F574F9A3E}
+		{046F33B0-9573-4AFA-9A68-9B16344F9AF7} = {8F8A6A34-59A8-446F-A994-A98F574F9A3E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {BA90D36C-5292-4D69-A21F-53581CFC6B5E}
