+++ src/Chirp.CSVDBService/Chirp.CSVDBService.csproj
@@ -7,7 +7,7 @@
   </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\Chirp.CLI\Chirp.CLI.csproj" />
+        <ProjectReference Include="..\Chirp.Core\Chirp.Core.csproj" />
     </ItemGroup>
 
 </Project>
