+++ src/Chirp.Core/Cheep.cs
@@ -1,4 +1,4 @@
-﻿namespace Chirp.CLI;
+﻿namespace Chirp.Core;
 
 public record Cheep(string Author, string Message, long Timestamp) 
 {
